+++ include/metalang99/maybe.h
@@ -61,16 +61,16 @@
  *
  * @code
  * #include <metalang99/maybe.h>
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
- * M_maybeEq(v(M_uintEq), M_just(v(123)), M_just(v(123)));
+ * M_maybeEq(v(M_natEq), M_just(v(123)), M_just(v(123)));
  *
  * // 0
- * M_maybeEq(v(M_uintEq), M_just(v(4)), M_just(v(6)));
+ * M_maybeEq(v(M_natEq), M_just(v(4)), M_just(v(6)));
  *
  * // 0
- * M_maybeEq(v(M_uintEq), M_just(v(4)), M_nothing());
+ * M_maybeEq(v(M_natEq), M_just(v(4)), M_nothing());
  * @endcode
  */
 #define METALANG99_maybeEq(compare, maybe, other)                                                  \