+++ include/metalang99/nat.h
@@ -1,17 +1,15 @@
 /**
  * @file
- * Unsigned integers.
- *
- * Currently, unsigned integers can range only from 0 to 255, inclusively.
+ * Natural numbers ([0; 255]).
  */
 
-#ifndef METALANG99_UINT_H
-#define METALANG99_UINT_H
+#ifndef METALANG99_NAT_H
+#define METALANG99_NAT_H
 
-#include <metalang99/uint/dec.h>
-#include <metalang99/uint/div.h>
-#include <metalang99/uint/eq.h>
-#include <metalang99/uint/inc.h>
+#include <metalang99/nat/dec.h>
+#include <metalang99/nat/div.h>
+#include <metalang99/nat/eq.h>
+#include <metalang99/nat/inc.h>
 
 #include <metalang99/control.h>
 #include <metalang99/lang.h>
@@ -25,7 +23,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 6
  * M_inc(v(5))
@@ -41,7 +39,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 4
  * M_dec(v(5))
@@ -57,46 +55,46 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * #define MATCH_Z_IMPL()  v(Billie)
  * #define MATCH_S_IMPL(x) v(Jean ~ x)
  *
  * // Billie
- * M_uintMatch(v(0), v(MATCH_))
+ * M_natMatch(v(0), v(MATCH_))
  *
  * // Jean ~ 122
- * M_uintMatch(v(123), v(MATCH_))
+ * M_natMatch(v(123), v(MATCH_))
  * @endcode
  *
  * @note This function calls @p f with #METALANG99_call, so no partial application occurs, and so
  * arity specifiers are not needed.
  */
-#define METALANG99_uintMatch(x, matcher) METALANG99_call(METALANG99_uintMatch, x, matcher)
+#define METALANG99_natMatch(x, matcher) METALANG99_call(METALANG99_natMatch, x, matcher)
 
 /**
- * The same as #METALANG99_uintMatch but provides additional arguments to all branches.
+ * The same as #METALANG99_natMatch but provides additional arguments to all branches.
  *
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * #define MATCH_Z_IMPL(x, y, z)    v(Billie ~ x y z)
  * #define MATCH_S_IMPL(n, x, y, z) v(Jean ~ n ~ x y z)
  *
  * // Billie ~ 1 2 3
- * M_uintMatchWithArgs(v(0), v(MATCH_), v(1, 2, 3))
+ * M_natMatchWithArgs(v(0), v(MATCH_), v(1, 2, 3))
  *
  * // Jean ~ 122 ~ 1 2 3
- * M_uintMatchWithArgs(v(123), v(MATCH_), v(1, 2, 3))
+ * M_natMatchWithArgs(v(123), v(MATCH_), v(1, 2, 3))
  * @endcode
  *
  * @note This function calls @p f with #METALANG99_call, so no partial application occurs, and so
  * arity specifiers are not needed.
  */
-#define METALANG99_uintMatchWithArgs(x, matcher, ...)                                              \
-    METALANG99_call(METALANG99_uintMatchWithArgs, x, matcher, __VA_ARGS__)
+#define METALANG99_natMatchWithArgs(x, matcher, ...)                                               \
+    METALANG99_call(METALANG99_natMatchWithArgs, x, matcher, __VA_ARGS__)
 
 /**
  * \f$x = y\f$
@@ -104,16 +102,16 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
- * M_uintEq(v(5), v(5))
+ * M_natEq(v(5), v(5))
  *
  * // 0
- * M_uintEq(v(3), v(8))
+ * M_natEq(v(3), v(8))
  * @endcode
  */
-#define METALANG99_uintEq(x, y) METALANG99_call(METALANG99_uintEq, x, y)
+#define METALANG99_natEq(x, y) METALANG99_call(METALANG99_natEq, x, y)
 
 /**
  * \f$x \neq y\f$
@@ -121,16 +119,16 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 0
- * M_uintNeq(v(5), v(5))
+ * M_natNeq(v(5), v(5))
  *
  * // 1
- * M_uintNeq(v(3), v(8))
+ * M_natNeq(v(3), v(8))
  * @endcode
  */
-#define METALANG99_uintNeq(x, y) METALANG99_call(METALANG99_uintNeq, x, y)
+#define METALANG99_natNeq(x, y) METALANG99_call(METALANG99_natNeq, x, y)
 
 /**
  * \f$x > y\f$
@@ -138,7 +136,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
  * M_greater(v(8), v(3))
@@ -155,7 +153,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
  * M_greaterEq(v(8), v(8))
@@ -172,7 +170,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
  * M_lesser(v(3), v(8))
@@ -189,7 +187,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 1
  * M_lesserEq(v(8), v(8))
@@ -206,7 +204,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 11
  * M_add(v(5), v(6))
@@ -220,7 +218,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 6
  * M_sub(v(11), v(5))
@@ -234,7 +232,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 12
  * M_mul(v(3), v(4))
@@ -248,7 +246,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 3
  * M_div(v(12), v(4))
@@ -265,7 +263,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // M_just(3)
  * M_divChecked(v(12), v(4))
@@ -285,7 +283,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 2
  * M_mod(v(8), v(3))
@@ -301,7 +299,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 15
  * M_add3(v(1), v(6), v(8))
@@ -315,7 +313,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 3
  * M_sub3(v(8), v(2), v(3))
@@ -329,7 +327,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 24
  * M_mul3(v(2), v(3), v(4))
@@ -343,7 +341,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 5
  * M_div(v(30), v(3), v(2))
@@ -359,7 +357,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 5
  * M_min(v(5), v(7))
@@ -373,7 +371,7 @@
  * # Examples
  *
  * @code
- * #include <metalang99/uint.h>
+ * #include <metalang99/nat.h>
  *
  * // 7
  * M_max(v(5), v(7))
@@ -392,54 +390,54 @@
 #define METALANG99_decPlain(x) METALANG99_PRIV_dec(x)
 
 /**
- * The plain version of #METALANG99_uintEq.
+ * The plain version of #METALANG99_natEq.
  */
-#define METALANG99_uintEqPlain(x, y) METALANG99_PRIV_uintEq(x, y)
+#define METALANG99_natEqPlain(x, y) METALANG99_PRIV_natEq(x, y)
 
 /**
- * The plain version of #METALANG99_uintNeq.
+ * The plain version of #METALANG99_natNeq.
  */
-#define METALANG99_uintNeqPlain(x, y) METALANG99_notPlain(METALANG99_uintEqPlain(x, y))
+#define METALANG99_natNeqPlain(x, y) METALANG99_notPlain(METALANG99_natEqPlain(x, y))
 // }
 
 #ifndef DOXYGEN_IGNORE
 
 // Implementation {
-#define METALANG99_uintMatch_IMPL(x, matcher)                                                      \
+#define METALANG99_natMatch_IMPL(x, matcher)                                                       \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(x, 0),                                                              \
+        METALANG99_PRIV_natEq(x, 0),                                                               \
         METALANG99_callTrivial(matcher##Z, ),                                                      \
         METALANG99_callTrivial(matcher##S, METALANG99_PRIV_dec(x)))
 
-#define METALANG99_uintMatchWithArgs_IMPL(x, matcher, ...)                                         \
+#define METALANG99_natMatchWithArgs_IMPL(x, matcher, ...)                                          \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(x, 0),                                                              \
+        METALANG99_PRIV_natEq(x, 0),                                                               \
         METALANG99_callTrivial(matcher##Z, __VA_ARGS__),                                           \
         METALANG99_callTrivial(matcher##S, METALANG99_PRIV_dec(x), __VA_ARGS__))
 
 #define METALANG99_inc_IMPL(x) v(METALANG99_incPlain(x))
 #define METALANG99_dec_IMPL(x) v(METALANG99_decPlain(x))
 
-#define METALANG99_uintEq_IMPL(x, y)  v(METALANG99_uintEqPlain(x, y))
-#define METALANG99_uintNeq_IMPL(x, y) v(METALANG99_uintNeqPlain(x, y))
+#define METALANG99_natEq_IMPL(x, y)  v(METALANG99_natEqPlain(x, y))
+#define METALANG99_natNeq_IMPL(x, y) v(METALANG99_natNeqPlain(x, y))
 
 #define METALANG99_greater_IMPL(x, y) METALANG99_lesser_IMPL(y, x)
 
 #define METALANG99_greaterEq_IMPL(x, y)                                                            \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(x, y),                                                              \
+        METALANG99_PRIV_natEq(x, y),                                                               \
         v(METALANG99_true),                                                                        \
         METALANG99_greater_IMPL(x, y))
 
 // METALANG99_lesser_IMPL {
 #define METALANG99_lesser_IMPL(x, y)                                                               \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_PRIV_natEq(y, 0),                                                               \
         v(METALANG99_false),                                                                       \
         METALANG99_callTrivial(METALANG99_PRIV_lesserProgress, x, METALANG99_PRIV_dec(y)))
 #define METALANG99_PRIV_lesserProgress_IMPL(x, y)                                                  \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(x, y),                                                              \
+        METALANG99_PRIV_natEq(x, y),                                                               \
         v(METALANG99_true),                                                                        \
         METALANG99_lesser_IMPL(x, y))
 // }
@@ -449,7 +447,7 @@
 // METALANG99_add_IMPL {
 #define METALANG99_add_IMPL(x, y)                                                                  \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_PRIV_natEq(y, 0),                                                               \
         v(x),                                                                                      \
         METALANG99_callTrivial(METALANG99_PRIV_addProgress, x, y))
 #define METALANG99_PRIV_addProgress_IMPL(x, y)                                                     \
@@ -459,7 +457,7 @@
 // METALANG99_sub_IMPL {
 #define METALANG99_sub_IMPL(x, y)                                                                  \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_PRIV_natEq(y, 0),                                                               \
         v(x),                                                                                      \
         METALANG99_callTrivial(METALANG99_PRIV_subProgress, x, y))
 #define METALANG99_PRIV_subProgress_IMPL(x, y)                                                     \
@@ -469,7 +467,7 @@
 // METALANG99_mul_IMPL {
 #define METALANG99_mul_IMPL(x, y)                                                                  \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_PRIV_natEq(y, 0),                                                               \
         v(0),                                                                                      \
         METALANG99_callTrivial(METALANG99_PRIV_mulProgress, x, y))
 #define METALANG99_PRIV_mulProgress_IMPL(x, y)                                                     \
@@ -479,7 +477,7 @@
 // METALANG99_mod_IMPL {
 #define METALANG99_mod_IMPL(x, y)                                                                  \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(y, 0),                                                              \
+        METALANG99_PRIV_natEq(y, 0),                                                               \
         METALANG99_fatal(METALANG99_mod, modulo by 0),                                             \
         METALANG99_callTrivial(METALANG99_PRIV_modAux, x, y, 0))
 
@@ -493,7 +491,7 @@
 
 #define METALANG99_PRIV_modProgress_IMPL(x, y, acc)                                                \
     METALANG99_PRIV_IF(                                                                            \
-        METALANG99_PRIV_uintEq(x, 0),                                                              \
+        METALANG99_PRIV_natEq(x, 0),                                                               \
         v(acc),                                                                                    \
         METALANG99_callTrivial(METALANG99_PRIV_modProgressAux, x, y, acc))
 
@@ -513,28 +511,28 @@
 // }
 
 // Arity specifiers {
-#define METALANG99_inc_ARITY               1
-#define METALANG99_dec_ARITY               1
-#define METALANG99_uintMatch_ARITY         2
-#define METALANG99_uintMatchWithArgs_ARITY 3
-#define METALANG99_uintEq_ARITY            2
-#define METALANG99_uintNeq_ARITY           2
-#define METALANG99_greater_ARITY           2
-#define METALANG99_greaterEq_ARITY         2
-#define METALANG99_lesser_ARITY            2
-#define METALANG99_lesserEq_ARITY          2
-#define METALANG99_add_ARITY               2
-#define METALANG99_sub_ARITY               2
-#define METALANG99_mul_ARITY               2
-#define METALANG99_div_ARITY               2
-#define METALANG99_divChecked_ARITY        2
-#define METALANG99_mod_ARITY               2
-#define METALANG99_add3_ARITY              3
-#define METALANG99_sub3_ARITY              3
-#define METALANG99_mul3_ARITY              3
-#define METALANG99_div3_ARITY              3
-#define METALANG99_min_ARITY               2
-#define METALANG99_max_ARITY               2
+#define METALANG99_inc_ARITY              1
+#define METALANG99_dec_ARITY              1
+#define METALANG99_natMatch_ARITY         2
+#define METALANG99_natMatchWithArgs_ARITY 3
+#define METALANG99_natEq_ARITY            2
+#define METALANG99_natNeq_ARITY           2
+#define METALANG99_greater_ARITY          2
+#define METALANG99_greaterEq_ARITY        2
+#define METALANG99_lesser_ARITY           2
+#define METALANG99_lesserEq_ARITY         2
+#define METALANG99_add_ARITY              2
+#define METALANG99_sub_ARITY              2
+#define METALANG99_mul_ARITY              2
+#define METALANG99_div_ARITY              2
+#define METALANG99_divChecked_ARITY       2
+#define METALANG99_mod_ARITY              2
+#define METALANG99_add3_ARITY             3
+#define METALANG99_sub3_ARITY             3
+#define METALANG99_mul3_ARITY             3
+#define METALANG99_div3_ARITY             3
+#define METALANG99_min_ARITY              2
+#define METALANG99_max_ARITY              2
 
 #define METALANG99_PRIV_modProgress_ARITY 3
 // }
@@ -542,37 +540,37 @@
 // Aliases {
 #ifndef METALANG99_NO_SMALL_PREFIX
 
-#define M_inc               METALANG99_inc
-#define M_dec               METALANG99_dec
-#define M_uintMatch         METALANG99_uintMatch
-#define M_uintMatchWithArgs METALANG99_uintMatchWithArgs
-#define M_uintEq            METALANG99_uintEq
-#define M_uintNeq           METALANG99_uintNeq
-#define M_greater           METALANG99_greater
-#define M_greaterEq         METALANG99_greaterEq
-#define M_lesser            METALANG99_lesser
-#define M_lesserEq          METALANG99_lesserEq
-#define M_add               METALANG99_add
-#define M_sub               METALANG99_sub
-#define M_mul               METALANG99_mul
-#define M_div               METALANG99_div
-#define M_divChecked        METALANG99_divChecked
-#define M_mod               METALANG99_mod
-#define M_add3              METALANG99_add3
-#define M_sub3              METALANG99_sub3
-#define M_mul3              METALANG99_mul3
-#define M_div3              METALANG99_div3
-#define M_min               METALANG99_min
-#define M_max               METALANG99_max
+#define M_inc              METALANG99_inc
+#define M_dec              METALANG99_dec
+#define M_natMatch         METALANG99_natMatch
+#define M_natMatchWithArgs METALANG99_natMatchWithArgs
+#define M_natEq            METALANG99_natEq
+#define M_natNeq           METALANG99_natNeq
+#define M_greater          METALANG99_greater
+#define M_greaterEq        METALANG99_greaterEq
+#define M_lesser           METALANG99_lesser
+#define M_lesserEq         METALANG99_lesserEq
+#define M_add              METALANG99_add
+#define M_sub              METALANG99_sub
+#define M_mul              METALANG99_mul
+#define M_div              METALANG99_div
+#define M_divChecked       METALANG99_divChecked
+#define M_mod              METALANG99_mod
+#define M_add3             METALANG99_add3
+#define M_sub3             METALANG99_sub3
+#define M_mul3             METALANG99_mul3
+#define M_div3             METALANG99_div3
+#define M_min              METALANG99_min
+#define M_max              METALANG99_max
 
-#define M_incPlain     METALANG99_incPlain
-#define M_decPlain     METALANG99_decPlain
-#define M_uintEqPlain  METALANG99_uintEqPlain
-#define M_uintNeqPlain METALANG99_uintNeqPlain
+#define M_incPlain    METALANG99_incPlain
+#define M_decPlain    METALANG99_decPlain
+#define M_natEqPlain  METALANG99_natEqPlain
+#define M_natNeqPlain METALANG99_natNeqPlain
 
 #endif // METALANG99_NO_SMALL_PREFIX
 // }
 
 #endif // DOXYGEN_IGNORE
 
-#endif // METALANG99_UINT_H
+#endif // METALANG99_NAT_H